improvement

+++ MasterChess/src/main/java/com/utc/api13/client/ihm/controllers/ModifyProfileController.java
@@ -169,19 +169,21 @@ public class ModifyProfileController {
         fileChooser.setTitle("Ouvrir le document");
         fileChooser.setInitialDirectory(new File("/"));
         File f = fileChooser.showOpenDialog(new Stage());
-        try {
-            String extensionFile = FilenameUtils.getExtension(f.getAbsolutePath());
-            File dest = Paths.get("user/avatar_"+ UUID.randomUUID().toString() +"."+extensionFile).toFile();
-            copyFile(f, dest);
-            changeProfilePicture.setImage(new Image("file:///" + dest.getAbsolutePath()));
-            dest.getAbsolutePath();
-        } catch (Exception e) {
-            try {
-                error("Error when changing your picture", false);
-            } catch (IOException e1) {
-                LOGGER.error(e1.getMessage(), e1);
-            }
-            LOGGER.error(e.getMessage(), e);
+        if (f !=null && f.exists()){
+	        try {
+	            String extensionFile = FilenameUtils.getExtension(f.getAbsolutePath());
+	            File dest = Paths.get("user/avatar_"+ UUID.randomUUID().toString() +"."+extensionFile).toFile();
+	            copyFile(f, dest);
+	            changeProfilePicture.setImage(new Image("file:///" + dest.getAbsolutePath()));
+	            dest.getAbsolutePath();
+	        } catch (Exception e) {
+	            try {
+	                error("Error when changing your picture", false);
+	            } catch (IOException e1) {
+	                LOGGER.error(e1.getMessage(), e1);
+	            }
+	            LOGGER.error(e.getMessage(), e);
+	        }
         }
 
     }
